Add javadocs,update gradle build and update README
