+++ build.gradle
@@ -31,6 +31,68 @@ dependencies {
     compile group: 'com.fasterxml.jackson.core', name: 'jackson-databind', version:'2.9.6'
     compile group: 'com.beust', name: 'jcommander', version: jCommanderVersion
     testCompile group: 'org.hsqldb', name: 'hsqldb', version:'2.4.0'
+    testCompile group: 'junit', name: 'junit', version: '4.12'
+
 }
 
+task run (type: JavaExec, dependsOn: classes, overwrite: true){
+    if(project.hasProperty('appArgs')){
+        args(appArgs.split(','))
+    }
+    description = "FCM XMPP App server"
+    main = mainClassName
+    classpath = sourceSets.main.runtimeClasspath
+}
 
+//create a single Jar with all dependencies
+task fatJar(type: Jar) {
+    dependsOn configurations.runtime
+    manifest {
+        attributes 'Implementation-Title': 'radar-fcmxmppserver',
+                'Implementation-Version': version,
+                'Main-Class': mainClassName
+    }
+    baseName = project.name + '-all'
+    from {
+        configurations.runtime.collect {
+            it.isDirectory() ? it : zipTree(it)
+        }
+    } {
+        exclude 'META-INF/*'
+    }
+    with jar
+}
+
+artifacts {
+    archives fatJar
+}
+
+distributions {
+    main {
+        contents {
+            into ("share/${project.name}") {
+                from 'README.md', 'LICENSE-2.0.txt'
+            }
+        }
+    }
+}
+
+tasks.withType(Tar){
+    compression = Compression.GZIP
+    extension = 'tar.gz'
+}
+
+task downloadDependencies {
+    description "Pre-downloads dependencies"
+    configurations.compileClasspath.files
+    configurations.runtimeClasspath.files
+}
+
+task copyDependencies(type: Copy) {
+    from configurations.runtimeClasspath.files
+    into "${buildDir}/third-party/"
+}
+
+wrapper {
+    gradleVersion '4.8'
+}
