+++ settings.gradle
@@ -1,1 +1,1 @@
-rootProject.name = 'xmppserver2'
+rootProject.name = 'radar-xmppserver'
