+++ src/main/java/org/radarcns/xmppserver/commandline/PathValidator.java
@@ -3,6 +3,11 @@ package org.radarcns.xmppserver.commandline;
 import com.beust.jcommander.IParameterValidator;
 import com.beust.jcommander.ParameterException;
 
+/**
+ * Path validator for command line args for the XMPP Server
+ *
+ * @author yatharthranjan
+ */
 public class PathValidator implements IParameterValidator{
     @Override
     public void validate(String name, String value) throws ParameterException {
