+++ src/main/java/org/radarcns/xmppserver/config/Config.java
@@ -1,0 +1,14 @@
+package org.radarcns.xmppserver.config;
+
+/**
+ * Config class for the scheduler in XMPP Server
+ *
+ * @author yatharthranjan
+ */
+public class Config {
+
+    public static final String SCHEDULER_SIMPLE = "simple";
+    public static final String SCHEDULER_MEM = "in-memory";
+    public static final String SCHEDULER_PERSISTENT = "persistent";
+
+}
