+++ src/main/java/org/radarcns/xmppserver/model/Data.java
@@ -4,6 +4,8 @@ package org.radarcns.xmppserver.model;
 /**
  *
  * Model class for storing data payload of messages sent downstream.
+ *
+ * @author yatharth
  */
 public class Data {
     // TODO add data fields
